refactor(async): give the deferred output its own module
AsyncRBLNModelRunnerOutput shares nothing with the runner but its constructor
arguments: it holds what the step handed it and copies it on another thread.
rbln_model_runner.py is over three thousand lines, so a self-contained eighty
of them are better read on their own.
